Discord/requirementsd.txt
discord.py==2.3.2
sqlite3
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openai

